Clobbered block with block2.
